Separate test classes to avoid flaky testing failures
